Fixed bug
fixed : call serializer.py and models.py in appropriate way  in every views file
(
from users.models import User
from users.serializers import UserSerializer
)

fixed: rearrange urls for index in project level
