+++ Backend/Backend/urls.py
@@ -17,16 +17,15 @@ Including another URLconf
 
 from django.contrib import admin
 from django.urls import path, include
+from users import views
 
 urlpatterns = [
     
+    #path('', views.index),
     
-    path("admin/", admin.site.urls),
-    
-    path('users/', include('users.urls')), # Include the urls from the users app
+    path('', include('users.urls')), # Include the urls from the users app
     path('admins/', include('admins.urls')), # Include the urls from the admins app
     
-    
-
+    path("admin/", admin.site.urls),
     
 ]
