--- Backend/admins/views.py/__init__.py
@@ -1,3 +1,0 @@
-
-from posts import *
-from postRequests import *
