+++ Backend/admins/views/posts.py
@@ -1,6 +1,6 @@
 from django.http import JsonResponse
-from models import Post
-from serializers import PostSerializer
+from admins.models import Post
+from admins.serializers import PostSerializer
 
 from rest_framework.decorators import api_view 
 from rest_framework.response import Response
