+++ Backend/users/views/users.py
@@ -1,6 +1,7 @@
+
 from django.http import JsonResponse
-from models import User
-from serializers import UserSerializer
+from users.models import User
+from users.serializers import UserSerializer
 
 from rest_framework.decorators import api_view 
 from rest_framework.response import Response
